+++ packages/dart/lib/src/math.dart
@@ -42,3 +42,13 @@ num mul(List<num> args) {
   return total;
 }
 
+/// Returns after subtracting all n arguments of numbers or the values of a single array of numbers.
+num sub(List<num> args) {
+  num total = args[0];
+
+  for (var i = 1; i < args.length; i++) {
+    total -= args[i];
+  }
+
+  return total;
+}