+++ packages/dart/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog (Dart)
 
-## 1.1.12 (2026--)
+## 1.1.12 (2026-03-31)
 
 - **BREAKING CHANGES**: `numRandom` has renamed to `numPick`
 - `getFileName`: Fix incorrect directory name with include dot character